Resolve line break issue for special-vowels
This change addresses the problem of unexpected line breaks being
inserted before non-consonant characters. The issue was due to a
missing check in the isUnicodeConsonant function which only considered
half of the vowels. The update adds the missing check by including the
check for the special vowels, ensuring that isUnicodeConsonant
correctly returns true only for actual consonant letters.